requirements.txt
gradio
tensorflow-cpu==2.5.0
numpy
requests